+++ releng/build_and_test.sh
@@ -70,6 +70,7 @@ echo " --"
 echo ""
 
 
+(cd $DIR && ./releng/fetch-rcptt-runner.sh)
 time (cd $DIR && mvn -e -c ${BATCHMODE} clean ${BUILDGOAL} ${TESTMODE})
 
 # Sonar